requirements.txt
python-dotenv
playwright
requests
Flask